Move to OpenSSH (sshd)
This changes srvcfg-manager service to support the OpenSSH server
instead of the Dropbear SSH server.  That is, it creates a D-Bus
object called "sshd" instead of "dropbear".

This change affects all users of the "dropbear" object, for example,
if bmcweb hard-coded "dropbear".
The community is exploring a better way to adapt to the choice of SSH
server implementation.

Tested: Yes, via bmcweb

Signed-off-by: Joseph Reynolds <joseph.reynolds1@ibm.com>